# public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ghost Protocol API",
    "version": "1.0.0",
    "description": "Machine-readable API surface for GhostRank discovery and GhostGate access."
  },
  "servers": [
    {
      "url": "https://ghostprotocol.cc"
    }
  ],
  "paths": {
...
    "/api/agents": {
      "get": {
        "summary": "List indexed agents",
        "operationId": "listAgents",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000,
              "default": 100
            }
          },
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "q",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "owner",
            "in": "query",
            "schema": {
              "type": "string",
              "pattern": "^0x[a-fA-F0-9]{40}$"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "volume"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent list payload"
          },
          "400": {
            "description": "Invalid query parameters"
          }
        }
      }
    },
...
  }
}